+++ test/unit/replication.test.ts
@@ -68,7 +68,6 @@ import { RxDBAttachmentsPlugin } from '../../plugins/attachments/index.mjs';
 import { RxDBMigrationSchemaPlugin } from '../../plugins/migration-schema/index.mjs';
 import { RxDBCleanupPlugin } from '../../plugins/cleanup/index.mjs';
 
-
 type CheckpointType = any;
 type TestDocType = HumanWithTimestampDocumentType;
 
@@ -1297,7 +1296,77 @@ describe('replication.test.ts', () => {
             sub2.unsubscribe();
             await dbV2.close();
         });
+        it('#7264 Replication pause ensureNotFalsy() throws', async () => {
+            // create a schema
+            const mySchema = {
+                version: 0,
+                primaryKey: 'passportId',
+                type: 'object',
+                properties: {
+                    passportId: {
+                        type: 'string',
+                        maxLength: 100,
+                    },
+                    firstName: {
+                        type: 'string',
+                    },
+                    lastName: {
+                        type: 'string',
+                    },
+                    age: {
+                        type: 'integer',
+                        minimum: 0,
+                        maximum: 150,
+                    },
+                },
+            };
 
+            /**
+             * Always generate a random database-name
+             * to ensure that different test runs do not affect each other.
+             */
+            const name = randomToken(10);
+
+            // create a database
+            const db = await createRxDatabase({
+                name,
+                storage: config.storage.getStorage()
+            });
+            // create a collection
+            const collections = await db.addCollections({
+                mycollection: {
+                    schema: mySchema,
+                },
+            });
+            const replicationState = replicateRxCollection({
+                collection: collections.mycollection,
+                replicationIdentifier: 'my-collection-http-replication',
+                waitForLeadership: false,
+                live: true,
+                pull: {
+                    handler: async () => {
+                        await wait(0);
+                        return {
+                            checkpoint: null,
+                            documents: [],
+                        };
+                    },
+                },
+                push: {
+                    handler: async () => {
+                        await wait(0);
+                        return [];
+                    },
+                },
+            });
+
+            await replicationState.pause();
+
+            db.close();
+            replicationState.cancel();
+
+
+        });
         it('#7187 real-time query ignoring the latest changes after deleting and purging data', async () => {
             if (
                 config.storage.name.includes('random-delay') ||